fix(app): show what a publisher decided, and preview the same verdict
Address PR review - the surfaces a publisher's review reaches.

The app never showed an author-dismissed finding, which made disclosure —
the entire justification for honouring somebody else's judgement — a claim
only the CLI kept. A row carrying findings is no longer sorted by its
verdict: settling every finding on an item makes it install and leaves the
findings there to read, so "clean" now means nothing was found rather than
nothing counts. Under Checked, one disclosure lists each of them with its
severity, rule, location, message, and the publisher, reason and date
recorded with the record; a row with open findings beside settled ones
lists both. A marketplace package's page marks a settled finding the same
way, in the publisher's name.

Browse was the fourth "how safe is this" path and the only one still
ignoring the record: a reviewed package showed as held back on the page a
person reads right before installing, then installed at 100. It now calls
the same derivation the gate, the audit and the authoring check use, with
the record read beside the thresholds rather than baked into the cache, so
a review recorded after the score was cached takes effect without
re-scoring and the cache format is unchanged.

The reusable workflow's strict guard was always true: in a called workflow
the github context is the caller's, so event_name holds their trigger and
never workflow_call, and every external catalog was being forced --strict
and made to run kendex's own real-CLI suite against their checkout. Both
now ask inputs, the discriminator the file already used for ref, with the
reason written down so it does not come back.

The safety print indexes decisions rather than zipping them, so a finding
whose decision is missing says so instead of vanishing from the one surface
whose job is to show every finding. The ARCHITECTURE invariant names the
third decision class, where its record lives, that it is read before the
verdict and can therefore unblock unlike the person's, and how the three
compose.

+++ crates/core/src/quality/author.rs
@@ -36,20 +36,23 @@ use crate::model::ItemKind;
 use crate::source_read::SealedSource;
 
 /// One finding the publisher settled, and how far that reaches.
+// Written into the lock, which is JSON in camelCase throughout; the
+// authoring file's own kebab-case shape is `reviews::Dismissal`, and these
+// are two different records that happen to rhyme.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "kebab-case")]
+#[serde(rename_all = "camelCase")]
 pub struct AuthorDismissal {
     pub reason: DismissReason,
     pub dismissed_at: String,
     /// How many times the publisher's own bytes carried this finding. The
     /// budget a reader spends: past it, an occurrence is content the
     /// publisher never reviewed.
-    pub occurrences: usize,
+    pub occurrences: u32,
 }
 
 /// A publisher's decisions about one item, as they travel to an install.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "kebab-case")]
+#[serde(rename_all = "camelCase")]
 pub struct AuthorReview {
     /// The bytes this speaks for: the source content at plan time, the
     /// installed content once the lock records it.
@@ -96,7 +99,7 @@ impl AuthorReview {
 /// How many occurrences of each finding are already settled. Empty settles
 /// nothing, which is what every failure here falls back to.
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct Budget(BTreeMap<String, usize>);
+pub struct Budget(BTreeMap<String, u32>);
 
 impl Budget {
     /// A decision made against these very bytes, which therefore speaks for
@@ -105,7 +108,7 @@ impl Budget {
         Budget(
             fingerprints
                 .into_iter()
-                .map(|fingerprint| (fingerprint, usize::MAX))
+                .map(|fingerprint| (fingerprint, u32::MAX))
                 .collect(),
         )
     }
@@ -283,7 +286,7 @@ fn source_occurrences(
     kind: ItemKind,
     name: &str,
     item_path: &Path,
-) -> Result<BTreeMap<String, usize>> {
+) -> Result<BTreeMap<String, u32>> {
     let file = item_path
         .strip_prefix(sealed.root())
         .unwrap_or(item_path)
@@ -296,7 +299,7 @@ fn source_occurrences(
         location: file.clone(),
         content: content(sealed, kind, item_path)?,
     });
-    let mut counts: BTreeMap<String, usize> = BTreeMap::new();
+    let mut counts: BTreeMap<String, u32> = BTreeMap::new();
     for finding in &result.findings {
         *counts.entry(finding.fingerprint(&file)).or_default() += 1;
     }
